Bubbled up the PromotePrecision
