+++ sql-plugin/src/main/311until330-nondb/scala/com/nvidia/spark/rapids/shims/Spark31Xuntil33XShims.scala
@@ -22,7 +22,7 @@ import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.execution._
 import org.apache.spark.sql.execution.datasources.v2._
 
-trait Spark31Xuntil33XShims extends SparkShims {
+trait Spark31Xuntil33XShims extends SparkBaseShim {
 
   def neverReplaceShowCurrentNamespaceCommand: ExecRule[_ <: SparkPlan] = {
     GpuOverrides.neverReplaceExec[ShowCurrentNamespaceExec]("Namespace metadata operation")
