+++ sql-plugin/src/main/320+-nondb/scala/com/nvidia/spark/rapids/shims/Spark320PlusNonDBShims.scala
@@ -15,7 +15,6 @@
  */
 package com.nvidia.spark.rapids.shims
 
-import com.nvidia.spark.rapids.SparkShims
 import org.apache.hadoop.fs.FileStatus
 
 import org.apache.spark.sql.catalyst.InternalRow
@@ -30,7 +29,7 @@ import org.apache.spark.sql.execution.python.WindowInPandasExec
 /**
  * Shim methods that can be compiled with every supported 3.2.0+ except Databricks versions
  */
-trait Spark320PlusNonDBShims extends SparkShims {
+trait Spark320PlusNonDBShims extends SparkBaseShim {
 
   override final def broadcastModeTransform(mode: BroadcastMode, rows: Array[InternalRow]): Any =
     mode.transform(rows)
